Move LabeledRange
