+++ packages/palette/src/components/index.tsx
@@ -1,2 +1,3 @@
 export * from "./EntityHeader"
+export * from "./LabeledRange"
 export * from "./Menu"
