+++ packages/palette/src/elements/index.tsx
@@ -14,7 +14,6 @@ export * from "./Input"
 export * from "./Join"
 export * from "./Message"
 export * from "./Pagination"
-export * from "./LabeledRange"
 export * from "./Link"
 export * from "./PriceRange"
 export * from "./ProgressBar"
